feat: backend feature + tests
