+++ iaso/api/bulk_create_users/utils.py
@@ -1,4 +1,4 @@
-def detect_multi_field_value_splitter(dialect, value):
+def detect_multi_field_value_splitter_csv(dialect, value):
     delimiter = dialect.delimiter
     alternatives = {
         ";": [",", "|"],
@@ -14,3 +14,11 @@ def detect_multi_field_value_splitter(dialect, value):
             return alt
 
     return delimiter
+
+
+def detect_multi_field_value_splitter_xls(value):
+    possible_delimiters = [";", "|", ","]
+    counts = {char: value.count(char) for char in possible_delimiters}
+    most_present_delimiter = max(counts, key=counts.get)
+
+    return most_present_delimiter
