+++ iaso/tests/api/bulk_create_users/test_views/test_create_from_csv.py
@@ -1,50 +1,23 @@
 import csv
 import io
-import json
 
 from unittest import mock
-from unittest.mock import patch
 
-import jsonschema
-
-from django.contrib.auth import get_user_model
-from django.contrib.auth.models import Permission, User
+from django.contrib.auth.models import User
 from django.core.files.uploadedfile import SimpleUploadedFile
 from django.urls import reverse
 from rest_framework import status
 
-from iaso import models as m
-from iaso.models import BulkCreateUserCsvFile, Profile, Team, UserRole
-from iaso.permissions.core_permissions import (
-    CORE_USERS_ADMIN_PERMISSION,
-    CORE_USERS_MANAGED_PERMISSION,
-)
-from iaso.tests.api.bulk_create_users.test_views.common import BulkCreateBaseAPITestCase
-from iaso.tests.api.profiles.test_views.common import PROFILE_LOG_SCHEMA
+from iaso.models import Profile
+from iaso.test import APITestCase
+from iaso.tests.api.bulk_create_users.test_views.common import BulkCreateAPITestCase
 
 
 BASE_URL = "/api/bulkcreateuser/"
 
 
-class BulkCreateFromCsvTestCase(BulkCreateBaseAPITestCase):
-    CSV_HEADER = [
-        "username",
-        "password",
-        "email",
-        "first_name",
-        "last_name",
-        "orgunit",
-        "orgunit__source_ref",
-        "profile_language",
-        "dhis2_id",
-        "permissions",
-        "user_roles",
-        "projects",
-        "teams",
-        "phone_number",
-        "organization",
-        "editable_org_unit_types",
-    ]
+class BulkCreateFromCsvTestCase(BulkCreateAPITestCase, APITestCase):
+    _file_extension = "csv"
 
     def setUp(self):
         super().setUp()
@@ -53,430 +26,29 @@ class BulkCreateFromCsvTestCase(BulkCreateBaseAPITestCase):
         self.mock_from_buffer = patcher.start()
         self.addCleanup(patcher.stop)
 
-    def test_upload_valid_csv(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
+    def create_from_fixture(self, fixture_name, context):
+        return self.load_fixture_with_jinja_template(
             path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid.csv",
+            fixture_name=f"{fixture_name}.csv",
             context=context,
         )
 
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
+    def generate_simple_uploaded_file(self, file_name, file_content):
+        return SimpleUploadedFile(f"{file_name}.csv", file_content.encode("utf-8"), content_type="text/csv")
 
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        users = User.objects.all()
-        profiles = Profile.objects.all()
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(len(users), 8)
-        self.assertEqual(len(profiles), 8)
-        new_user_1 = users.get(username="broly")
-        org_unit_ids = [org_unit.id for org_unit in list(new_user_1.iaso_profile.org_units.all())]
-        self.assertEqual(new_user_1.email, "biobroly@bluesquarehub.com")
-        self.assertEqual(new_user_1.first_name, "broly")
-        self.assertEqual(new_user_1.last_name, "bio")
-        self.assertEqual(new_user_1.iaso_profile.language, "fr")
-        self.assertEqual(new_user_1.iaso_profile.dhis2_id, "dhis2_id_1")
-        self.assertEqual(org_unit_ids, [9999])
-
-        # Checking that the file was uploaded to the right location
-        file_upload = m.BulkCreateUserCsvFile.objects.first()
-        # The API endpoint changes the file name after processing - doing this with file.save() generates a random suffix
-        # This means that it's impossible to check for strict equality
-        expected_file_name = f"{self.account1.short_sanitized_name}_{self.account1.id}/bulk_create_user_csv/{file_upload.created_at.strftime('%Y_%m')}/test_user_bulk_create_valid.csv"
-        self.assertEqual(file_upload.file.name, expected_file_name)
-
-        # Verify default options are stored as empty when not provided
-        self.assertEqual(file_upload.default_profile_language, "")
-        self.assertEqual(file_upload.default_organization, "")
-        self.assertEqual(file_upload.default_permissions.count(), 0)
-        self.assertEqual(file_upload.default_user_roles.count(), 0)
-        self.assertEqual(file_upload.default_projects.count(), 0)
-        self.assertEqual(file_upload.default_org_units.count(), 0)
-        self.assertEqual(file_upload.default_teams.count(), 0)
-
-        # check that the file passwords are masked
-        file_upload.file.seek(0)  # Ensure pointer is at start
-        csv_content = io.StringIO(file_upload.file.read().decode("utf-8"))
+    def get_rows_from_file(self, file):
+        csv_content = io.StringIO(file.read().decode("utf-8"))
         reader = csv.DictReader(csv_content)
-        for row in reader:
-            self.assertEqual(row["password"], "******")
 
-    def test_num_queries(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
-
-        with self.assertNumQueries(47):
-            response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-    def test_num_queries_with_perms(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        self.account1.modules = self.MODULES
-        self.account1.save()
-
-        self.account1.refresh_from_db()
-        with self.assertNumQueries(47):
-            with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_valid_with_perm.csv") as csv_users:
-                response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-    def test_upload_valid_csv_with_perms(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        iaso_forms = Permission.objects.get(codename="iaso_forms")
-        iaso_submissions = Permission.objects.get(codename="iaso_submissions")
-
-        self.account1.modules = self.MODULES
-        self.account1.save()
-
-        self.account1.refresh_from_db()
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_valid_with_perm.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        pollux = User.objects.get(username="pollux")
-        pollux_perms = pollux.user_permissions.all()
-        has_perms = False
-        if iaso_forms and iaso_submissions in pollux_perms:
-            has_perms = True
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(has_perms, True)
-
-    def test_upload_invalid_mail_csv(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_mail.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        self.assertEqual(len(validation_errors), 1)
-        self.assertEqual(int(validation_errors[0]["row"]), 3)
-        self.assertIn("email", validation_errors[0]["details"])
-
-    def test_upload_without_mail_must_work(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_no_mail.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertJSONResponse(response, 201)
-        usernames = ["broly", "ferdinand", "rsfg"]
-
-        self.assertEqual(get_user_model().objects.filter(username__in=usernames).count(), 3)
-
-    def test_upload_invalid_orgunit_id(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_orgunit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        # Row 1 has invalid orgunit ID, rows 2-3 have invalid source_ref
-        self.assertEqual(len(validation_errors), 3)
-
-        # Check first error - invalid orgunit ID
-        self.assertEqual(int(validation_errors[0]["row"]), 1)
-        self.assertIn("orgunit", validation_errors[0]["details"])
-        self.assertIn("99998", validation_errors[0]["details"]["orgunit"][0])
-
-        # Check second and third errors - invalid source_ref
-        self.assertEqual(int(validation_errors[1]["row"]), 2)
-        self.assertIn("orgunit__source_ref", validation_errors[1]["details"])
-        self.assertEqual(int(validation_errors[2]["row"]), 3)
-        self.assertIn("orgunit__source_ref", validation_errors[2]["details"])
-
-    def test_upload_user_already_exists(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        user = User.objects.create(
-            username="broly", first_name="broly", last_name="Smith", email="broly-smith@bluesquarehub.com"
-        )
-        user.save()
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_orgunit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        # Should have error for existing username
-        username_error = next((e for e in validation_errors if "username" in e.get("details", {})), None)
-        self.assertIsNotNone(username_error)
-        self.assertEqual(username_error["details"]["username"], ["This field must be unique."])
-
-    def test_upload_invalid_csv_dont_create_entries(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_orgunit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        users = User.objects.all()
-        profiles = Profile.objects.all()
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        self.assertTrue(len(validation_errors) > 0)
-        self.assertEqual(len(users), 5)
-        self.assertEqual(len(profiles), 5)
-
-    def test_can_create_user_with_managed_geo_limit_permission(self):
-        self.client.force_authenticate(self.user_managed_geo_limit)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_managed_geo_limit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-    def test_cant_create_user_not_org_unit_in_the_pyramid_with_managed_geo_limit_permission(self):
-        self.client.force_authenticate(self.user_managed_geo_limit)
-        self.source.projects.set([self.project])
-        self.user_managed_geo_limit.iaso_profile.org_units.add(self.org_unit1)
-        self.user_managed_geo_limit.save()
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_managed_geo_limit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        for ve in validation_errors:
-            self.assertEqual(ve["details"], {"orgunit": ["Object with pk=1112 does not exist."]})
-
-    def test_user_cant_access_without_permission(self):
-        self.client.force_authenticate(self.obi)
-        self.source.projects.set([self.project])
-
-        response = self.client.get(f"{BASE_URL}")
-
-        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
-
-    def test_user_can_access_with_managed_geo_limit_permission(self):
-        self.client.force_authenticate(self.user_managed_geo_limit)
-        self.source.projects.set([self.project])
-
-        response = self.client.post(f"{BASE_URL}")
-
-        self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-
-    def test_password_delete_after_import(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        pswd_deleted = True
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
-
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        csv_file = BulkCreateUserCsvFile.objects.last()
-
-        reader = csv.reader(csv_file.file.open("r"))
-        i = 0
-        csv_indexes = []
-        for row in reader:
-            if i > 0:
-                pswd_deleted = True if row[csv_indexes.index("password")] == "" else False
-            else:
-                csv_indexes = row
-            i += 1
-
-        self.assertEqual(pswd_deleted, True)
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-    def test_created_users_can_login(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        login_data = {"username": "broly", "password": "yodnj!30dln"}
-
-        login_response = self.client.post("/api/token/", data=login_data, format="json")
-
-        self.assertEqual(login_response.status_code, 200)
-
-    def test_upload_invalid_orgunit_name(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_ou_name.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        orgunit_error = next((e for e in validation_errors if "orgunit" in e.get("details", {})), None)
-        self.assertIsNotNone(orgunit_error)
-        self.assertIn("Bazarre", orgunit_error["details"]["orgunit"][0])
-
-    def test_users_profiles_have_right_ou(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        broly_ou = Profile.objects.get(user=User.objects.get(username="broly").id).org_units.all()
-        ou_list = []
-
-        for ou in broly_ou:
-            ou_list.append(ou.id)
-
-        ferdinand_ou = Profile.objects.get(user=User.objects.get(username="ferdinand").id).org_units.all()
-        ou_f_list = []
-
-        for ou in ferdinand_ou:
-            ou_f_list.append(ou.id)
-
-        self.assertEqual(ou_list, [9999])
-        self.assertCountEqual(ou_f_list, [self.org_unit1.id, self.org_unit2.id, 9999])
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-    def test_cant_create_user_without_access_to_ou(self):
-        self.client.force_authenticate(self.yoda)
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-
-    def test_cant_create_user_without_ou_profile(self):
-        self.client.force_authenticate(self.yoda)
-
-        with open(
-            "iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_creator_no_access_to_ou.csv"
-        ) as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        orgunit_error = next((e for e in validation_errors if "orgunit" in e.get("details", {})), None)
-        self.assertIsNotNone(orgunit_error)
-        # The error should indicate access issue or invalid org unit
-        self.assertIn("10244", orgunit_error["details"]["orgunit"][0])
-
-    def test_upload_duplicate_ou_names(self):
-        # This test detects if in case there is multiple OU with same name in different sources the right OU is taken.
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open(
-            "iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_user_duplicate_ou_names.csv"
-        ) as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertEqual(User.objects.filter(username="jan").exists(), True)
-
-        created_user = User.objects.get(username="jan")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(len(created_user.iaso_profile.org_units.all()), 1)
-        self.assertEqual(created_user.iaso_profile.org_units.all()[0].name, "chiloe")
-        self.assertEqual(created_user.iaso_profile.org_units.all()[0].id, 10244)
-
-    def test_can_create_user_with_ou_that_are_child_of_ou(self):
-        # This test ensure that we can create users with access to ou child not explicitly added to the creator
-        # access ou list ( ou is a child of ou in the user ou list )
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open(
-            "iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_user_access_to_child_ou.csv"
-        ) as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(User.objects.filter(username="jan").exists(), True)
-
-        created_user = User.objects.get(username="jan")
-
-        self.assertEqual(len(created_user.iaso_profile.org_units.all()), 1)
-        self.assertEqual(created_user.iaso_profile.org_units.all()[0].name, "chiloe")
-        self.assertEqual(created_user.iaso_profile.org_units.all()[0].id, 10244)
+        yield from reader
 
     def test_upload_semicolon_separated_csv(self):
         self.client.force_authenticate(self.yoda)
         self.source.projects.set([self.project])
 
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_semicolon.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
+        file = self.generate_file_from_fixture("test_user_bulk_create_semicolon", with_context=False)
+
+        response = self.client.post(f"{BASE_URL}", {"file": file}, format="multipart")
 
         users = User.objects.all()
         profiles = Profile.objects.all()
@@ -497,588 +69,6 @@ class BulkCreateFromCsvTestCase(BulkCreateBaseAPITestCase):
         self.assertEqual(new_user_1.iaso_profile.dhis2_id, "dhis2_id_1")
         self.assertEqual(new_user_2.iaso_profile.dhis2_id, "dhis2_id_6")
         self.assertEqual(org_unit_ids, [9999])
-
-    def test_upload_csv_with_missing_column(self):
-        self.client.force_authenticate(self.yoda)
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_missing_columns.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content", {"general": "Missing required column(s): permissions"})
-
-    def test_create_user_with_roles(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        manager = {"name": "manager"}
-        self.client.post("/api/userroles/", data=manager)
-
-        area_manager = {"name": "area_manager"}
-        self.client.post("/api/userroles/", data=area_manager)
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_valid_with_roles.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        users = User.objects.all()
-        profiles = Profile.objects.all()
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(len(users), 7)
-        self.assertEqual(len(profiles), 7)
-        new_user_1 = users.get(username="broly")
-        new_user_2 = users.get(username="cyrus")
-        new_user_1_user_role = new_user_1.iaso_profile.user_roles.all()
-        new_user_2_user_roles = new_user_2.iaso_profile.user_roles.all()
-        org_unit_ids = [org_unit.id for org_unit in list(new_user_1.iaso_profile.org_units.all())]
-        self.assertEqual(new_user_1.email, "biobroly@bluesquarehub.com")
-        self.assertEqual(new_user_2.email, "cyruswashington@bluesquarehub.com")
-        self.assertEqual(new_user_1.first_name, "broly")
-        self.assertEqual(new_user_1.last_name, "bio")
-        self.assertEqual(new_user_2.first_name, "cyrus")
-        self.assertEqual(new_user_2.last_name, "washington")
-        self.assertEqual(new_user_1.iaso_profile.language, "fr")
-        self.assertEqual(new_user_1.iaso_profile.dhis2_id, "dhis2_id_1")
-        self.assertEqual(new_user_2.iaso_profile.dhis2_id, "dhis2_id_6")
-        self.assertEqual(org_unit_ids, [9999])
-        self.assertEqual(len(new_user_1_user_role), 1)
-        self.assertEqual(len(new_user_1.groups.all()), 1)
-        self.assertEqual(new_user_1.groups.all().first(), new_user_1_user_role.first().group)
-        self.assertEqual(len(new_user_2_user_roles), 2)
-        self.assertEqual(len(new_user_2.groups.all()), 2)
-        self.assertCountEqual(
-            [user_role.group for user_role in new_user_2_user_roles],
-            [group for group in new_user_2.groups.all()],
-        )
-
-    def test_create_user_with_projects(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_valid_with_projects.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        users = User.objects.all()
-        profiles = Profile.objects.all()
-        profile_1 = Profile.objects.get(user__username="broly")
-        profile_2 = Profile.objects.get(user__username="cyrus")
-
-        self.assertEqual(profile_1.projects.all()[0].name, "Project name")
-        self.assertEqual(profile_2.projects.all()[0].name, "Project name")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(len(users), 7)
-        self.assertEqual(len(profiles), 7)
-        new_user_1 = users.get(username="broly")
-        new_user_2 = users.get(username="cyrus")
-        org_unit_ids = [org_unit.id for org_unit in list(new_user_1.iaso_profile.org_units.all())]
-        self.assertEqual(new_user_1.email, "biobroly@bluesquarehub.com")
-        self.assertEqual(new_user_2.email, "cyruswashington@bluesquarehub.com")
-        self.assertEqual(new_user_1.first_name, "broly")
-        self.assertEqual(new_user_1.last_name, "bio")
-        self.assertEqual(new_user_2.first_name, "cyrus")
-        self.assertEqual(new_user_2.last_name, "washington")
-        self.assertEqual(new_user_1.iaso_profile.language, "fr")
-        self.assertEqual(new_user_1.iaso_profile.dhis2_id, "dhis2_id_1")
-        self.assertEqual(new_user_2.iaso_profile.dhis2_id, "dhis2_id_6")
-        self.assertEqual(org_unit_ids, [9999])
-
-    def test_create_user_with_project_restrictions(self):
-        self.source.projects.set([self.project, self.project2])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_managed_geo_limit.csv") as csv_users:
-            csv_reader = list(csv.reader(csv_users))
-
-            csv_line_1 = csv_reader[1]
-            csv_line_2 = csv_reader[2]
-            csv_line_3 = csv_reader[3]
-
-            username_index = 0
-            projects_index = 12
-
-            username_1 = csv_line_1[username_index]
-            username_2 = csv_line_2[username_index]
-            username_3 = csv_line_3[username_index]
-
-            # Ensure the CSV tries to set `project` as an authorized project.
-            self.assertEqual(csv_line_1[projects_index], self.project.name)
-            self.assertEqual(csv_line_2[projects_index], self.project.name)
-
-        self.user_managed_geo_limit.iaso_profile.projects.set([self.project2.id])  # Restrict user to `project2`.
-        # Grant access to org_unit_child (1112) which is used in the CSV
-        self.user_managed_geo_limit.iaso_profile.org_units.add(self.org_unit_child)
-        self.assertTrue(self.user_managed_geo_limit.has_perm(CORE_USERS_MANAGED_PERMISSION.full_name()))
-        self.assertFalse(self.user_managed_geo_limit.has_perm(CORE_USERS_ADMIN_PERMISSION.full_name()))
-
-        self.client.force_authenticate(self.user_managed_geo_limit)
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_managed_geo_limit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        # changing this as it made no sense. If the project passed is invalid , we just put the available projects instead and silence the error
-        # on the other hand, passing empty project without default made it so it stayed empty. inconsistent
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-
-        self.assertEqual(len(validation_errors), 2)
-
-        self.assertIn("projects", validation_errors[0]["details"])
-        self.assertIn("Object with name=Project name does not exist.", validation_errors[0]["details"]["projects"][0])
-
-        self.client.force_authenticate(self.yoda)
-        self.yoda.iaso_profile.org_units.add(self.org_unit_child)
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_managed_geo_limit.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        profile_1 = Profile.objects.get(user__username=username_1)
-        self.assertEqual(1, profile_1.projects.count())
-        self.assertEqual(profile_1.projects.first(), self.project)
-        profile_2 = Profile.objects.get(user__username=username_2)
-        self.assertEqual(1, profile_2.projects.count())
-        self.assertEqual(profile_2.projects.first(), self.project)
-        profile_3 = Profile.objects.get(user__username=username_3)
-        self.assertEqual(0, profile_3.projects.count())
-
-    def test_should_create_user_with_the_correct_org_unit_from_source_ref(self):
-        """
-        Given that multiple instances of an OrgUnit exist, when attempting to link
-        an OrgUnit based on its `source_ref` to a new user, then the OrgUnit that
-        matches the "default source version" of the account should be selected.
-        """
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        org_unit_a = m.OrgUnit.objects.create(name="Zakuul", source_ref="qux", version=self.version1)
-        org_unit_b = m.OrgUnit.objects.create(name="Zakuul", source_ref="qux", version=self.version2)
-        org_unit_c = m.OrgUnit.objects.create(name="Zakuul", source_ref="qux", version=None)
-        self.yoda.iaso_profile.org_units.add(org_unit_a, org_unit_b, org_unit_c)
-        self.yoda.iaso_profile.save()
-
-        csv_str = io.StringIO()
-        writer = csv.DictWriter(csv_str, fieldnames=self.CSV_HEADER)
-        writer.writeheader()
-        writer.writerow(
-            {
-                "username": "john",
-                "password": "yodnj!30dln",
-                "email": "john@foo.com",
-                "first_name": "John",
-                "last_name": "Doe",
-                "orgunit": "",
-                "orgunit__source_ref": "qux",
-                "profile_language": "fr",
-                "dhis2_id": "",
-                "permissions": "",
-                "user_roles": "",
-                "projects": "",
-                "teams": "",
-                "phone_number": "",
-                "organization": "",
-                "editable_org_unit_types": "",
-            }
-        )
-        csv_bytes = csv_str.getvalue().encode()
-        csv_file = SimpleUploadedFile("users.csv", csv_bytes)
-
-        response = self.client.post(f"{BASE_URL}", {"file": csv_file}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        new_user = User.objects.get(email="john@foo.com")
-        self.assertEqual(new_user.iaso_profile.account, self.account1)
-        self.assertEqual(new_user.iaso_profile.org_units.count(), 1)
-        self.assertEqual(new_user.iaso_profile.org_units.first(), org_unit_a)
-        self.assertEqual(org_unit_a.version_id, self.account1.default_version_id)
-
-    def test_audit_log_on_save(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        self.account1.modules = self.MODULES
-        self.account1.save()
-        self.account1.refresh_from_db()
-        # Manually adding user roles to avoid messing with other tests
-        manager = {"name": "manager"}
-        self.client.post("/api/userroles/", data=manager)
-
-        area_manager = {"name": "area_manager"}
-        self.client.post("/api/userroles/", data=area_manager)
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_all_fields.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        self.client.force_authenticate(self.john)
-        response = self.client.get("/api/logs/?contentType=iaso.profile&fields=past_value,new_value")
-        response_data = self.assertJSONResponse(response, status.HTTP_200_OK)
-        logs = response_data["list"]
-        logs_for_new_users = [logs[0], logs[1], logs[2]]  # There's 3 users in the csv so we take the last 3 logs
-
-        # Check that all users are logged
-        user_names = [log["new_value"][0]["fields"]["username"] for log in logs_for_new_users]
-        self.assertIn("bob", user_names)
-        self.assertIn("bobette", user_names)
-        self.assertIn("fanchon", user_names)
-
-        # Check that fields have correct value (1 user)
-        log = logs[0]
-
-        try:
-            jsonschema.validate(instance=log, schema=PROFILE_LOG_SCHEMA)
-        except jsonschema.exceptions.ValidationError as ex:
-            self.fail(msg=str(ex))
-
-        past_value = log["past_value"]
-        self.assertEqual(past_value, [])
-
-        new_value = log["new_value"][0]["fields"]
-        self.assertTrue(new_value["password_updated"])
-        self.assertNotIn("password", new_value.keys())
-        self.assertEqual(len(new_value["user_permissions"]), 1)
-        self.assertIn("iaso_forms", new_value["user_permissions"])
-        self.assertEqual(len(new_value["user_roles"]), 1)
-        self.assertEqual(len(new_value["projects"]), 1)
-        self.assertEqual(new_value["language"], "fr")
-        self.assertEqual(new_value["dhis2_id"], "dhis2_id_3")
-        self.assertEqual(len(new_value["org_units"]), 2)
-        self.assertIn(self.org_unit3.id, new_value["org_units"])
-        self.assertIn(self.org_unit2.id, new_value["org_units"])
-
-    def test_bulk_create_user_with_single_team(self):
-        self.client.force_authenticate(self.yoda)
-
-        self.source.projects.set([self.project])
-
-        team = Team.objects.create(name="Alpha Team", project=self.project, manager=self.yoda)
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id, "team_names": "Alpha Team"}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid_with_multiple_teams.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile("test.csv", csv_content.encode("utf-8"), content_type="text/csv")
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        user = User.objects.get(username="rsfg")
-        self.assertEqual(user.teams.count(), 1)
-        self.assertEqual(user.teams.first().name, team.name)
-
-    def test_bulk_create_user_with_multiple_teams(self):
-        self.client.force_authenticate(self.yoda)
-
-        self.source.projects.set([self.project])
-
-        team1 = Team.objects.create(name="Alpha Team", project=self.project, manager=self.yoda)
-        team2 = Team.objects.create(name="Beta Team", project=self.project, manager=self.yoda)
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id, "team_names": f"{team1.name}, {team2.name}"}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid_with_multiple_teams.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile("test.csv", csv_content.encode("utf-8"), content_type="text/csv")
-
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        user = User.objects.get(username="rsfg")
-        self.assertEqual(user.teams.count(), 2)
-
-        user_team_names = list(user.teams.values_list("name", flat=True))
-        self.assertIn("Alpha Team", user_team_names)
-        self.assertIn("Beta Team", user_team_names)
-
-    def test_bulk_create_user_fails_on_invalid_team(self):
-        self.client.force_authenticate(self.yoda)
-
-        self.source.projects.set([self.project])
-        invalid_team_name = "Not a Team"
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id, "team_names": invalid_team_name}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid_with_multiple_teams.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile("test.csv", csv_content.encode("utf-8"), content_type="text/csv")
-
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-
-        self.assertHasError(res_data, "file_content")
-
-        validation_errors = res_data["file_content"]
-        self.assertEqual(len(validation_errors), 1)
-        self.assertIn("teams", validation_errors[0]["details"])
-        self.assertIn(invalid_team_name, validation_errors[0]["details"]["teams"][0])
-
-    def test_bulk_create_user_cannot_assign_team_from_another_account(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        account2 = m.Account.objects.get(name="Account 2")
-        han_solo = User.objects.get(username="han solo")
-
-        project_acc2 = m.Project.objects.create(name="Project Acc 2", account=account2, app_id="acc2.app")
-
-        secret_team_name = "Account 2 Team"
-        Team.objects.create(name=secret_team_name, project=project_acc2, manager=han_solo)
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id, "team_names": secret_team_name}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid_with_multiple_teams.csv",
-            context=context,
-        )
-
-        buffer = io.StringIO()
-        writer = csv.writer(buffer)
-
-        for row in csv_content.split("\n"):
-            writer.writerow(row.split(","))
-
-        csv_content = buffer.getvalue()
-
-        test_file = SimpleUploadedFile("test.csv", csv_content.encode("utf-8"), content_type="text/csv")
-
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-        validation_errors = res_data["file_content"]
-        self.assertEqual(len(validation_errors), 1)
-        self.assertIn("teams", validation_errors[0]["details"])
-        self.assertIn(secret_team_name, validation_errors[0]["details"]["teams"][0])
-
-    @patch("iaso.api.bulk_create_users.serializers.send_bulk_email_invitations")
-    def test_email_invitations_sent_for_users_without_password(self, mock_send_emails):
-        """Test that email invitations are sent only for users without passwords."""
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_email_no_password.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        # Verify that send_bulk_email_invitations was called
-        mock_send_emails.assert_called_once()
-        call_args = mock_send_emails.call_args[0]
-        user_ids = call_args[0]
-        self.assertEqual(len(user_ids), 2)
-
-        # Verify users were created correctly
-        invite1 = User.objects.get(username="invite1")
-        invite2 = User.objects.get(username="invite2")
-        self.assertFalse(invite1.has_usable_password())
-        self.assertFalse(invite2.has_usable_password())
-        self.assertEqual(invite1.email, "invite1@test.com")
-        self.assertEqual(invite2.email, "invite2@test.com")
-
-    @patch("iaso.api.bulk_create_users.serializers.send_bulk_email_invitations")
-    def test_email_invitations_mixed_scenario_with_fixture(self, mock_send_emails):
-        """Test email invitations in mixed scenario."""
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_mixed_email.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        # Verify that send_bulk_email_invitations was called
-        mock_send_emails.assert_called_once()
-        call_args = mock_send_emails.call_args[0]
-        user_ids = call_args[0]
-        self.assertEqual(len(user_ids), 2)
-
-        # Verify user creation and email/password states
-        with_password = User.objects.get(username="with_password")
-        no_password1 = User.objects.get(username="no_password1")
-        no_password2 = User.objects.get(username="no_password2")
-        no_email = User.objects.get(username="no_email")
-
-        self.assertTrue(with_password.has_usable_password())
-        self.assertTrue(no_email.has_usable_password())
-
-        self.assertFalse(no_password1.has_usable_password())
-        self.assertFalse(no_password2.has_usable_password())
-
-    def test_default_fields_three_scenarios(self):
-        """Test default field values in three scenarios: all applied, all ignored, and partial"""
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        self.account1.modules = self.MODULES
-        self.account1.save()
-        self.account1.refresh_from_db()
-
-        # Create teams for testing
-        bulk_team = Team.objects.create(name="Bulk Team", project=self.project, manager=self.yoda)
-        csv_team = Team.objects.create(name="Test Team", project=self.project, manager=self.yoda)
-
-        response = self.client.post("/api/userroles/", data={"name": "manager"})
-        manager_role = UserRole.objects.get(group__name=f"{self.account1.id}_manager")
-
-        # Get the permission ID for iaso_forms
-        iaso_forms_perm = Permission.objects.get(codename="iaso_forms")
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_bulk_configuration.csv") as csv_file:
-            response = self.client.post(
-                f"{BASE_URL}",
-                {
-                    "file": csv_file,
-                    "default_profile_language": "fr",
-                    "default_organization": "UNICEF",
-                    "default_permissions": json.dumps([iaso_forms_perm.id]),
-                    "default_user_roles": json.dumps([manager_role.id]),
-                    "default_projects": json.dumps([self.project.id]),
-                    "default_org_units": json.dumps([self.org_unit1.id, self.org_unit2.id]),
-                    "default_teams": json.dumps([bulk_team.id]),
-                },
-                format="multipart",
-            )
-
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        # Verify default options are stored in BulkCreateUserCsvFile (now M2M fields)
-        csv_file_record = BulkCreateUserCsvFile.objects.last()
-        self.assertEqual(csv_file_record.default_profile_language, "fr")
-        self.assertEqual(csv_file_record.default_organization, "UNICEF")
-        self.assertEqual(csv_file_record.default_permissions.count(), 1)
-        self.assertTrue(csv_file_record.default_permissions.filter(codename="iaso_forms").exists())
-        self.assertEqual(csv_file_record.default_user_roles.count(), 1)
-        self.assertEqual(csv_file_record.default_user_roles.first(), manager_role)
-        self.assertEqual(csv_file_record.default_projects.count(), 1)
-        self.assertEqual(csv_file_record.default_projects.first(), self.project)
-        self.assertEqual(csv_file_record.default_org_units.count(), 2)
-        self.assertIn(self.org_unit1, csv_file_record.default_org_units.all())
-        self.assertIn(self.org_unit2, csv_file_record.default_org_units.all())
-        self.assertEqual(csv_file_record.default_teams.count(), 1)
-        self.assertEqual(csv_file_record.default_teams.first(), bulk_team)
-
-        # ALL defaults applied (CSV fields empty)
-        all_bulk_user = User.objects.get(username="all_bulk_user")
-        all_bulk_profile = all_bulk_user.iaso_profile
-        self.assertEqual(all_bulk_profile.language, "fr")
-        self.assertEqual(all_bulk_profile.organization, "UNICEF")  # From defaults
-        self.assertIsNone(all_bulk_profile.dhis2_id)
-        self.assertTrue(all_bulk_user.user_permissions.filter(codename="iaso_forms").exists())
-        self.assertEqual(all_bulk_profile.user_roles.count(), 1)
-        self.assertEqual(all_bulk_profile.user_roles.first().group.name, f"{self.account1.id}_manager")
-        self.assertEqual(all_bulk_profile.projects.count(), 1)
-        self.assertEqual(all_bulk_profile.projects.first(), self.project)
-        self.assertEqual(all_bulk_profile.org_units.count(), 2)  # From defaults
-        self.assertIn(self.org_unit1, all_bulk_profile.org_units.all())
-        self.assertIn(self.org_unit2, all_bulk_profile.org_units.all())
-        self.assertEqual(all_bulk_user.teams.count(), 1)  # From defaults
-        self.assertEqual(all_bulk_user.teams.first().name, "Bulk Team")
-
-        # ALL defaults ignored (CSV fields have values)
-        no_bulk_user = User.objects.get(username="no_bulk_user")
-        no_bulk_profile = no_bulk_user.iaso_profile
-
-        self.assertEqual(no_bulk_profile.language, "en")
-        self.assertEqual(no_bulk_profile.organization, "MSF")
-        self.assertEqual(no_bulk_profile.dhis2_id, "csv_dhis2_1")
-        self.assertTrue(no_bulk_user.user_permissions.filter(codename="iaso_submissions").exists())
-        self.assertFalse(no_bulk_user.user_permissions.filter(codename="iaso_forms").exists())
-        self.assertEqual(no_bulk_profile.user_roles.count(), 1)
-        self.assertEqual(no_bulk_profile.user_roles.first().group.name, f"{self.account1.id}_manager")
-        self.assertEqual(no_bulk_profile.projects.count(), 1)
-        self.assertEqual(no_bulk_profile.projects.first().name, "Project 2")
-        self.assertEqual(no_bulk_profile.org_units.count(), 2)
-        self.assertIn(self.org_unit1, no_bulk_profile.org_units.all())
-        self.assertIn(self.org_unit2, no_bulk_profile.org_units.all())
-        self.assertEqual(no_bulk_user.teams.count(), 1)  # From CSV, not defaults
-        self.assertEqual(no_bulk_user.teams.first().name, "Test Team")
-
-        # PARTIAL defaults applied (some CSV fields empty)
-        mixed_user = User.objects.get(username="mixed_user")
-        mixed_profile = mixed_user.iaso_profile
-        self.assertEqual(mixed_profile.language, "en")
-        self.assertEqual(mixed_profile.organization, "WHO")
-        self.assertIsNone(mixed_profile.dhis2_id)
-        self.assertTrue(mixed_user.user_permissions.filter(codename="iaso_forms").exists())
-        self.assertFalse(mixed_user.user_permissions.filter(codename="iaso_submissions").exists())
-        self.assertEqual(mixed_profile.user_roles.count(), 1)
-        self.assertEqual(mixed_profile.user_roles.first().group.name, f"{self.account1.id}_manager")
-        self.assertEqual(mixed_profile.projects.count(), 1)
-        self.assertEqual(mixed_profile.projects.first(), self.project)
-        self.assertEqual(mixed_profile.org_units.count(), 2)
-        self.assertIn(self.org_unit1, mixed_profile.org_units.all())
-        self.assertIn(self.org_unit2, mixed_profile.org_units.all())
-        self.assertEqual(mixed_user.teams.count(), 1)  # From defaults (CSV empty)
-        self.assertEqual(mixed_user.teams.first().name, "Bulk Team")
-
-    def test_upload_invalid_permission_name(self):
-        """Test that CSV with invalid permission name returns validation error."""
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        self.account1.modules = self.MODULES
-        self.account1.save()
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_permission.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-        self.assertHasError(res_data, "file_content")
-
-        validation_errors = res_data["file_content"]
-        self.assertEqual(len(validation_errors), 1)
-        self.assertIn("permissions", validation_errors[0]["details"])
-        self.assertIn("iaso_fake_permission", validation_errors[0]["details"]["permissions"][0])
-
-    def test_upload_invalid_language(self):
-        """Test that CSV with invalid language returns validation error."""
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        with open("iaso/tests/fixtures/bulk_create_users/test_user_bulk_create_invalid_language.csv") as csv_users:
-            response = self.client.post(f"{BASE_URL}", {"file": csv_users}, format="multipart")
-
-        res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
-
-        self.assertHasError(res_data, "file_content")
-
-        self.assertEqual(len(res_data["file_content"]), 1)
-        self.assertIn("profile_language", res_data["file_content"][0]["details"])
-        self.assertIn("invalid_lang", res_data["file_content"][0]["details"]["profile_language"][0])
-
-    def test_create_user_with_project_by_id(self):
-        """Test that CSV with project ID works correctly."""
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        context = {"project_id": self.project.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_valid_with_project_ids.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile("test.csv", csv_content.encode("utf-8"), content_type="text/csv")
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-        res_json = self.assertJSONResponse(response, status.HTTP_201_CREATED)
-        self.assertEqual(res_json, {})
-
-        user = User.objects.get(username="projectid_user")
-        self.assertEqual(user.iaso_profile.projects.count(), 1)
-        self.assertEqual(user.iaso_profile.projects.first().id, self.project.id)
 
     def test_file_from_sample_does_not_trigger_invalid_file_type(self):
         self.client.force_authenticate(self.yoda)
@@ -1096,30 +86,3 @@ class BulkCreateFromCsvTestCase(BulkCreateBaseAPITestCase):
 
         res_data = self.assertJSONResponse(response, status.HTTP_400_BAD_REQUEST)
         self.assertNotIn("file", res_data)
-
-    def test_create_user_with_empty_first_name_and_last_name(self):
-        self.client.force_authenticate(self.yoda)
-        self.source.projects.set([self.project])
-
-        context = {"org_unit_type_id": self.org_unit_type_region.id}
-
-        csv_content = self.load_fixture_with_jinja_template(
-            path_to_fixtures="iaso/tests/fixtures/bulk_create_users",
-            fixture_name="test_user_bulk_create_empty_names.csv",
-            context=context,
-        )
-
-        test_file = SimpleUploadedFile(
-            "test_user_bulk_create_valid.csv", csv_content.encode("utf-8"), content_type="text/csv"
-        )
-
-        response = self.client.post(f"{BASE_URL}", {"file": test_file}, format="multipart")
-        self.assertJSONResponse(response, status.HTTP_201_CREATED)
-
-        users = User.objects.all()
-        self.assertEqual(users.count(), 6)
-
-        user = users.filter(username="broly").first()
-        self.assertEqual(user.first_name, "")
-        self.assertEqual(user.last_name, "")
-        self.assertEqual(user.email, "")
