+++ iaso/tests/fixtures/bulk_create_users/test_should_create_user_with_the_correct_org_unit_from_source_ref.csv
@@ -1,0 +1,2 @@
+username,password,email,first_name,last_name,orgunit,orgunit__source_ref,profile_language,dhis2_id,organization,permissions,user_roles,projects,teams,phone_number,editable_org_unit_types
+john,yodnj!30dln,john@foo.com,John,Doe,,qux,fr,,,,,,,,
